added verification propably on every endpoint to check if user is a creator of a wedding before starting doing any changes

+++ Backend/Backend.Application/Services/Wedding/WeddingService.cs
@@ -3,6 +3,7 @@ using Backend.Application.DTO.WeddingDTO;
 using Backend.Domain.Interfaces;
 using Backend.Domain.Entities;
 using QRCoder;
+using Microsoft.Extensions.Logging.Abstractions;
 
 
 namespace Backend.Application.Services.Wedding
@@ -68,37 +69,52 @@ namespace Backend.Application.Services.Wedding
             return weddingsDTO;
         }
 
-        public async Task<WeddingDetailsDTO> GetWeddingDetailsById(Guid id)
+        public async Task<WeddingDetailsDTO> GetWeddingDetailsById(Guid id, string userId)
         {
+            if (!await _weddingRepository.IsUserOwnerOfWedding(id, userId)) 
+            {
+                return null; 
+            }
             var wedding = await _weddingRepository.GetDetailsById(id);
             var weddingDetailsDTO = _mapper.Map<WeddingDetailsDTO>(wedding);
             return weddingDetailsDTO;
         }
 
 
-        public async Task<bool> Delete(Guid id)
+        public async Task<bool> Delete(Guid id, string userId)
         {
+            if (!await _weddingRepository.IsUserOwnerOfWedding(id, userId))
+            {
+                return false;
+            }
             return await _weddingRepository.DeleteWeedingById(id);
         }
 
 
-        public async Task<bool> Update(WeddingDTO newWeddingDTO)
+        public async Task<bool> Update(WeddingDTO newWeddingDTO, string userId)
         {
+            
+
             var oldWedding = await _weddingRepository.GetDetailsById(newWeddingDTO.Id);
 
-            if (oldWedding == null)
+            if (oldWedding == null || !await _weddingRepository.IsUserOwnerOfWedding(newWeddingDTO.Id, userId) )
             {
                 return false;
             }
 
             var updatedWedding = _mapper.Map(newWeddingDTO, oldWedding);
-
             var updateSuccess = await _weddingRepository.Update(oldWedding);
             return updateSuccess;
         }
 
-        public async Task<bool> ExtendSessionKeyExpiration(Guid Id, TimeSpan extensionDuration)
+        public async Task<bool> ExtendSessionKeyExpiration(Guid Id, TimeSpan extensionDuration, string userId)
         {
+            if (!await _weddingRepository.IsUserOwnerOfWedding(Id, userId))
+            {
+                return false;
+            }
+
+
             var wedding = await _weddingRepository.GetDetailsById(Id);
 
             if (wedding == null)
@@ -116,8 +132,14 @@ namespace Backend.Application.Services.Wedding
 
 
 
-        public async Task<byte[]> GetQrCode(Guid weddingId)
+        public async Task<byte[]> GetQrCode(Guid weddingId, string userId)
         {
+            if (!await _weddingRepository.IsUserOwnerOfWedding(weddingId, userId))
+            {
+                return null;
+            }
+
+
             var wedding = await _weddingRepository.GetDetailsById(weddingId);
             if (wedding.IsSessionKeyExpired || wedding.IsSessionKeyExpired )
             {
